add customer crud operations

+++ src/main/java/com/femas1/quickbill/service/AddressService.java
@@ -2,8 +2,10 @@ package com.femas1.quickbill.service;
 
 import com.femas1.quickbill.model.Address;
 import com.femas1.quickbill.model.Company;
+import com.femas1.quickbill.model.Customer;
 import com.femas1.quickbill.repository.AddressRepository;
 import com.femas1.quickbill.repository.CompanyRepository;
+import com.femas1.quickbill.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,23 +15,26 @@ import java.util.List;
 public class AddressService {
     private final AddressRepository addressRepository;
     private final CompanyRepository companyRepository;
+    private final CustomerRepository customerRepository;
 
     @Autowired
-    public AddressService(AddressRepository addressRepository, CompanyRepository companyRepository){
+    public AddressService(AddressRepository addressRepository, CompanyRepository companyRepository,
+                          CustomerRepository customerRepository){
         this.addressRepository = addressRepository;
         this.companyRepository = companyRepository;
+        this.customerRepository = customerRepository;
     }
 
     public Address createAddress(Address address){
         return addressRepository.save(address);
     }
 
-    public Address getAddress(Integer id){
+    public Address getAddress(int id){
         return addressRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Address not found."));
     }
 
-    public Address updateAddress(Integer id, Address address){
+    public Address updateAddress(int id, Address address){
         Address existingAddress = addressRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Address not found."));
 
@@ -45,10 +50,10 @@ public class AddressService {
             throw new RuntimeException("Address not found.");
         }
         List<Company> companiesUsingAddress = companyRepository.findByAddress_AddressId(id);
+        List<Customer> customersUsingAddress = customerRepository.findByAddress_AddressId(id);
 
-        if (!companiesUsingAddress.isEmpty()) {
-            throw new RuntimeException("Cannot delete address: It is still in use by "
-                    + companiesUsingAddress.size() + " companies.");
+        if (!companiesUsingAddress.isEmpty() || !customersUsingAddress.isEmpty()) {
+            throw new RuntimeException("Cannot delete address: It is still in use by a customer or company");
         }
         addressRepository.deleteById(id);
     }
